Warning fix In file included from /usr/include/c++/4.6/memory:76:0,                  from                  /home/dirvine/Devel/MaidSafe/src/common/include/maidsafe/common/test.h:32,                                   from                                   /home/dirvine/Devel/MaidSafe/src/lifestuff/src/maidsafe/lifestuff/tests/contacts_test.cc:15:                                   /home/dirvine/Devel/MaidSafe/src/lifestuff/src/maidsafe/lifestuff/tests/contacts_test.cc:                                   In member function ‘virtual void                                   maidsafe::lifestuff::test::ContactsTest_BEH_ListContacts_Status_Test::TestBody()’:                                   /home/dirvine/Devel/MaidSafe/src/lifestuff/src/maidsafe/lifestuff/tests/contacts_test.cc:218:3:                                   error: assuming signed overflow does                                   not occur when changing X +- C1 cmp C2                                   to X cmp C1 +- C2                                   [-Werror=strict-overflow]                                   /home/dirvine/Devel/MaidSafe/src/lifestuff/src/maidsafe/lifestuff/tests/contacts_test.cc:235:3:                                   error: assuming signed overflow does                                   not occur when changing X +- C1 cmp C2                                   to X cmp C1 +- C2                                   [-Werror=strict-overflow]                                   cc1plus: all warnings being treated as                                   errors

+++ src/maidsafe/lifestuff/tests/contacts_test.cc
@@ -206,8 +206,8 @@ TEST_F(ContactsTest, BEH_ListContacts_Rank_LastContact) {
 
 TEST_F(ContactsTest, BEH_ListContacts_Status) {
   Contact msc(contact_);
-  int indexing(0);
-  int status_index(0);
+  uint16_t indexing(0);
+  uint16_t status_index(0);
   std::vector<ContactStatus> types;
   types.push_back(kUninitialised);
   types.push_back(kRequestSent);
